image_sorter/requirements.txt
Pillow>=10.0.0
rich>=13.0.0